+++ .gitignore
@@ -1,54 +1,2 @@
-# ndk generated files
-**/obj/
-
-# built application files
-*.apk
-*.ap_
-
-# lint files
-lint.xml
-
-# files for the dex VM
-*.dex
-
-# Java class files
-*.class
-
-# generated files
-bin/
-gen/
-classes/
-gen-external-apklibs/
-
-# maven output folder
-target
-
-# Local configuration file (sdk path, etc)
-local.properties
-
-# Eclipse project files
-.classpath
-.project
-.metadata
-.settings
-
-# Proguard folder generated by Eclipse
-proguard/
-
-# IntelliJ files
-*.iml
-*.ipr
-*.iws
-.idea/
-
-# OSX files
-.DS_Store
-
-# Windows files
-Thumbs.db
-
-# vi swap files
-*.swp
-
-# backup files
-*.bak
+build
+.idea/workspace.xml
