+++ aFileChooser/src/main/AndroidManifest.xml
