+++ aFileChooser/src/main/res/drawable-hdpi/ic_chooser.png
