+++ aFileChooser/src/main/res/drawable-hdpi/ic_file.png
