+++ aFileChooser/src/main/res/drawable-hdpi/ic_folder.png
