+++ aFileChooser/src/main/res/drawable-hdpi/ic_provider.png
