+++ aFileChooser/src/main/res/drawable-mdpi/ic_chooser.png
