+++ aFileChooser/src/main/res/drawable-mdpi/ic_file.png
