+++ aFileChooser/src/main/res/drawable-mdpi/ic_folder.png
