+++ aFileChooser/src/main/res/drawable-mdpi/ic_provider.png
