+++ aFileChooser/src/main/res/drawable-xhdpi/ic_chooser.png
