+++ aFileChooser/src/main/res/drawable-xhdpi/ic_file.png
