+++ aFileChooser/src/main/res/drawable-xhdpi/ic_folder.png
