+++ aFileChooser/src/main/res/drawable-xhdpi/ic_provider.png
