+++ aFileChooser/src/main/res/drawable-xxhdpi/ic_chooser.png
