+++ aFileChooser/src/main/res/drawable-xxhdpi/ic_file.png
