+++ aFileChooser/src/main/res/drawable-xxhdpi/ic_folder.png
