+++ aFileChooser/src/main/res/drawable-xxhdpi/ic_provider.png
