+++ aFileChooser/src/main/res/values-de/strings.xml
