+++ aFileChooser/src/main/res/values-fr/strings.xml
