+++ aFileChooserExample/src/main/AndroidManifest.xml
