+++ aFileChooserExample/src/main/res/drawable/ic_launcher.png
