+++ aFileChooserExample/src/main/res/values/strings.xml
